Fix -Wconversion/-Wsign-conversion warnings in 9 small modules
Make integer conversions explicit with static_cast at the point of
conversion, and retype a few file-local variables (subsample loop
indices/abundances to uint64_t) where that removes the conversion at
its root. char subscripts go through unsigned char to stay
value-preserving. All converted values are non-negative lengths,
counts, indices, or abundances; no behavior change.

The pre-existing static_cast<int> narrowing of 64-bit abundances in
fasta2fastq.cc and fastq_join.cc is preserved verbatim (now written as
static_cast<uint64_t>(static_cast<int>(...))) and flagged for human
review -- it looks like a latent truncation bug, but fixing it would
change behavior.

+++ src/fasta2fastq.cc
@@ -111,7 +111,7 @@ auto fasta2fastq(struct Parameters const & parameters) -> void
                           fasta_get_header(fp_input),
                           static_cast<int>(fasta_get_header_length(fp_input)),
                           quality.data(),
-                          static_cast<int>(fastq_get_abundance(fp_input)),
+                          static_cast<uint64_t>(static_cast<int>(fastq_get_abundance(fp_input))),
                           counter,
                           -1.0);
